+++ src/functions/public/Out-PermissionFile.ps1
@@ -1,4 +1,4 @@
-function Out-PermissionReport {
+function Out-PermissionFile {
 
     # missing
 
@@ -50,7 +50,6 @@ function Out-PermissionReport {
         [Hashtable]$AceByGUID,
         [Hashtable]$AclByPath,
         [Hashtable]$PrincipalByID,
-        $BestPracticeIssue,
         [Hashtable]$Parent,
 
         <#
@@ -63,7 +62,7 @@ function Out-PermissionReport {
             5   Accounts with access                                                                        $PrincipalsByResolvedID.Values | %{$_} | Sort ResolvedAccountName
             6   Expanded resolved access rules (expanded with account info)                                 $Permissions
             7   Formatted permissions                                                                       $FormattedPermissions
-            8   Best Practice issues                                                                        $BestPracticeIssues
+            8   Best Practice issues                                                                        $BestPracticeEval
             9   XML custom sensor output for Paessler PRTG Network Monitor                                  $PrtgXml
             10  Permission Report
         #>
@@ -114,7 +113,7 @@ function Out-PermissionReport {
         [ValidateSet('none', 'all', 'target', 'item', 'account')]
         [string[]]$SplitBy = 'target',
 
-        [PSCustomObject]$BestPracticeIssues
+        [PSCustomObject]$BestPracticeEval
 
     )
 
@@ -139,7 +138,7 @@ function Out-PermissionReport {
 
     $UnsplitDetail = $Detail | Where-Object -FilterScript { $_ -le 5 -or $_ -in 8, 9 }
     $SplitDetail = $Detail | Where-Object -FilterScript { $_ -gt 5 -and $_ -notin 8, 9 }
-    pause
+
     $DetailScripts = @(
         { $TargetPath },
         { ForEach ($Key in $Parent.Keys) {
@@ -164,8 +163,8 @@ function Out-PermissionReport {
 
         },
         { $Permissions.Data },
-        { $BestPracticeIssues },
-        { $Permission.TargetPermissions },
+        { $BestPracticeEval },
+        { ConvertTo-PermissionPrtgXml -Analysis $Analysis },
         {}
     )
 
@@ -375,8 +374,6 @@ function Out-PermissionReport {
                 'prtgxml' {
 
                     $DetailExports = @( { }, { }, { }, { }, { }, { }, { }, { }, { }, { $args[0] | Out-File -LiteralPath $args[1] } )
-
-                    $DetailScripts[9] = { ConvertTo-PermissionPrtgXml -Analysis $Analysis }
                     break
 
                 }
@@ -482,19 +479,13 @@ function Out-PermissionReport {
 
                     }
 
-                    'prtgxml' {
-
-                        # Output the full path of the XML file (result of the custom XML sensor for Paessler PRTG Network Monitor) to the Information stream
-                        #Write-Information $XmlFile
-
-                        # Save the XML file to disk
-                        #$null = Set-Content -LiteralPath $XmlFile -Value $XMLOutput
-                        break
-
-                    }
+                    # Nothing for 'prtgxml' because it is an Unsplit detail level only
 
                     'xml' {
+
+                        Out-PermissionDetailReport -Detail $SplitDetail -ReportObject $ReportObjects -DetailExport $DetailExports -Format $Format -OutputDir $FormatDir -Culture $Culture -DetailString $DetailStrings
                         break
+
                     }
 
                 }